+++ Changelog.md
@@ -8,6 +8,7 @@ Compiler Features:
 
 
 Bugfixes:
+ * Type Checker: Disallow constructor of the same class to be used as modifier